% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{Clear}
\usage{
clr(...)
}
\description{
This function clears several optional areas of RStudio. If the function has
no arguments, it will by default clear the Global Environment, Plots
window, and the Console window.
}
